Added a method to sign arbitrary hashes.
